Mad Dog edits Sunday evening 27may2018:
- cleaned up plots in bivariate section.

- deleted section where I made a Treasury dummy variable.  If we somehow
manage to shrink the document to allow this back in, I'd like to re-add
it.  It demonstrates use of a dummy variable, which could earn us
points.

- Moved fraction of expenditure by party by coarsedescription to
secondary effects section.

- GLOBAL CHANGE TO ALL SECTIONS: edited beginning of R code blocks to
center plots and label each plot with a figure number and description.
NOTE: the numbering is NOT automatic.  I couldn't get that to work despite
following instructions.  This means that any further changes to plot
order (or plot additions/deletions) will impact the accuracy of other
plot labels and references to them in the text.